Add script to dump EBCDIC tables
The script goes through each supported single-byte and double-byte
EBCDIC CCSID supported by the system and dumps the entire thing as both
a Unicode mapping table similar to
https://www.unicode.org/Public/MAPPINGS/ISO8859/8859-1.TXT
as well as an HTML table, similar to the style that Wikipedia used to
have.

The script uses Python ctypes to ILE iconv APIs to convert each code
point in the source table to UTF-16. UTF-16 is chosen, since the ILE
iconv APIs support that encoding of Unicode the best. To determine which
CCSIDs are valid, it calls QTQGESP to determine the each CCSID's
encoding scheme. Encoding schemes are documented at
https://www.ibm.com/support/knowledgecenter/en/ssw_ibm_i_74/nls/rbagsccsidencodschemes.htm
and only 0x1100 (EBCDIC SBCS) and 0x1200 (EBCDIC DBCS) are dumped.